Refactor build task.
Split into build (web) and buildReactNative.
