+++ node-src/tasks/build.ts
@@ -1,15 +1,7 @@
 import { AnalyticsEvent } from '@cli/analytics/events';
 import * as Sentry from '@sentry/node';
-import {
-  createReadStream,
-  existsSync,
-  mkdirSync,
-  readFileSync,
-  renameSync,
-  type WriteStream,
-} from 'fs';
+import { readFileSync } from 'fs';
 import path from 'path';
-import { createInterface } from 'readline';
 import semver from 'semver';
 import tmp from 'tmp-promise';
 
@@ -19,58 +11,15 @@ import { isE2EBuild } from '../lib/e2eUtils';
 import { emailHash } from '../lib/emailHash';
 import { getPackageManagerRunCommand } from '../lib/getPackageManager';
 import { openLogFileStream } from '../lib/logFile';
-import { buildAndroid, buildIos } from '../lib/react-native/build';
-import { readExpoConfig } from '../lib/react-native/expoConfig';
-import { generateManifest } from '../lib/react-native/generateManifest';
 import { exitCodes, setExitCode } from '../lib/setExitCode';
 import { runCommand } from '../lib/shell/shell';
 import { createTask, transitionTo } from '../lib/tasks';
-import { Context, Task } from '../types';
+import { Context } from '../types';
 import { endActivity, startActivity } from '../ui/components/activity';
-import { buildFailed, reactNativeBuildFailed } from '../ui/messages/errors/buildFailed';
+import { buildFailed } from '../ui/messages/errors/buildFailed';
 import e2eBuildFailed from '../ui/messages/errors/e2eBuildFailed';
 import missingDependency from '../ui/messages/errors/missingDependency';
-import {
-  failed,
-  initial,
-  pending,
-  pendingAndroid,
-  pendingIOS,
-  pendingManifest,
-  skipped,
-  skippedForReactNative,
-  success,
-} from '../ui/tasks/build';
-
-const MAX_REACT_NATIVE_LOG_LINES = 20;
-
-async function readLastLines(filePath: string, lineCount: number): Promise<string> {
-  return new Promise((resolve, reject) => {
-    const buffer: string[] = [];
-    const rl = createInterface({ input: createReadStream(filePath), crlfDelay: Infinity });
-    rl.on('line', (line) => {
-      buffer.push(line);
-      if (buffer.length > lineCount) buffer.shift();
-    });
-    rl.on('close', () => resolve(buffer.join('\n')));
-    rl.on('error', reject);
-  });
-}
-
-export const setSourceDirectory = async (ctx: Context) => {
-  // if it has been set upstream of us, do not overwrite it
-  if (ctx.sourceDir) return;
-
-  if (ctx.options.outputDir) {
-    ctx.sourceDir = ctx.options.outputDir;
-  } else if (ctx.storybook && ctx.storybook.version && semver.lt(ctx.storybook.version, '5.0.0')) {
-    // Storybook v4 doesn't support absolute paths like tmp.dir would yield
-    ctx.sourceDir = 'storybook-static';
-  } else {
-    const temporaryDirectory = await tmp.dir({ unsafeCleanup: true, prefix: `chromatic-` });
-    ctx.sourceDir = temporaryDirectory.path;
-  }
-};
+import { failed, initial, pending, skipped, success } from '../ui/tasks/build';
 
 const isStatsFlagSupported = (ctx: Context) => {
   return ctx.storybook && ctx.storybook.version
@@ -87,12 +36,22 @@ const getStatsFlag = (ctx: Context) => {
     : '--webpack-stats-json';
 };
 
-export const setBuildCommand = async (ctx: Context) => {
-  // We don't currently support building React Native Storybook so we'll skip this for now
-  if (ctx.isReactNativeApp) {
-    return;
-  }
+export const setSourceDirectory = async (ctx: Context) => {
+  // if it has been set upstream of us, do not overwrite it
+  if (ctx.sourceDir) return;
 
+  if (ctx.options.outputDir) {
+    ctx.sourceDir = ctx.options.outputDir;
+  } else if (ctx.storybook && ctx.storybook.version && semver.lt(ctx.storybook.version, '5.0.0')) {
+    // Storybook v4 doesn't support absolute paths like tmp.dir would yield
+    ctx.sourceDir = 'storybook-static';
+  } else {
+    const temporaryDirectory = await tmp.dir({ unsafeCleanup: true, prefix: `chromatic-` });
+    ctx.sourceDir = temporaryDirectory.path;
+  }
+};
+
+export const setBuildCommand = async (ctx: Context) => {
   const buildCommand = ctx.flags?.buildCommand || ctx.options.buildCommand;
   const buildCommandOptions: string[] = [];
 
@@ -220,11 +179,6 @@ function trackBuildFailure(ctx: Context, errorCategory: string, err: any) {
 }
 
 export const buildStorybook = async (ctx: Context) => {
-  // We don't currently support building React Native projects so we'll skip this for now
-  if (ctx.isReactNativeApp) {
-    return;
-  }
-
   let logFile;
   if (ctx.options.storybookLogFile) {
     ctx.buildLogFile = path.resolve(ctx.options.storybookLogFile);
@@ -260,96 +214,6 @@ export const buildStorybook = async (ctx: Context) => {
   }
 };
 
-const runPlatformCommand = async (
-  ctx: Context,
-  platform: 'android' | 'ios',
-  command: string,
-  logStream: WriteStream
-) => {
-  ctx.log.debug('Running React Native build command:', command);
-  const label = platform === 'android' ? 'Android' : 'iOS';
-  logStream.write(`\n[chromatic] ${label} build: ${command}\n`);
-  try {
-    await runCommand(command, {
-      stdout: logStream,
-      stderr: logStream,
-    });
-  } catch (err) {
-    setExitCode(ctx, exitCodes.NPM_BUILD_STORYBOOK_FAILED, true);
-    throw new Error(`React Native build command failed: ${command}\n${err.message}`);
-  }
-};
-
-const resolvePlatforms = (ctx: Context) => {
-  return (ctx.announcedBuild?.browsers ?? []).filter(
-    (b): b is 'ios' | 'android' => b === 'ios' || b === 'android'
-  );
-};
-
-/* eslint-disable-next-line complexity, max-statements */
-export const buildReactNativeArtifacts = async (ctx: Context, task: Task) => {
-  // we skip this build step if not React Native, or we have prebuilt assets.
-  if (!ctx.isReactNativeApp || !!ctx.options.storybookBuildDir) return;
-
-  const platforms = resolvePlatforms(ctx);
-  const needsAndroid = platforms.includes('android');
-  const needsIos = platforms.includes('ios');
-
-  if (!needsAndroid && !needsIos) return;
-
-  mkdirSync(path.join(ctx.sourceDir, '.chromatic'), { recursive: true });
-  ctx.reactNativeBuildLogFile = path.join(ctx.sourceDir, '.chromatic', 'react-native-build.log');
-
-  const logStream = await openLogFileStream(ctx.reactNativeBuildLogFile);
-
-  try {
-    if (needsAndroid) {
-      transitionTo(pendingAndroid)(ctx, task);
-      const { androidBuildCommand } = ctx.options.reactNative ?? {};
-      ctx.log.debug({ androidBuildCommand }, 'Running Android build');
-      if (androidBuildCommand) {
-        await runPlatformCommand(ctx, 'android', androidBuildCommand, logStream);
-      } else {
-        const { artifactPath } = await buildAndroid(logStream);
-        renameSync(artifactPath, path.join(ctx.sourceDir, 'storybook.apk'));
-      }
-    }
-
-    if (needsIos) {
-      transitionTo(pendingIOS)(ctx, task);
-      const { iosBuildCommand } = ctx.options.reactNative ?? {};
-      ctx.log.debug({ iosBuildCommand }, 'Running iOS build');
-      if (iosBuildCommand) {
-        await runPlatformCommand(ctx, 'ios', iosBuildCommand, logStream);
-      } else {
-        const config = await readExpoConfig();
-        const { artifactPath } = await buildIos(config.name, logStream);
-        renameSync(artifactPath, path.join(ctx.sourceDir, 'storybook.app'));
-      }
-    }
-  } catch (buildError) {
-    setExitCode(ctx, exitCodes.NPM_BUILD_STORYBOOK_FAILED, true);
-    await new Promise<void>((resolve) => logStream.end(resolve));
-    const tail = await readLastLines(ctx.reactNativeBuildLogFile, MAX_REACT_NATIVE_LOG_LINES);
-    ctx.log.error(reactNativeBuildFailed(ctx, buildError, tail));
-    throw new Error(failed(ctx).output);
-  } finally {
-    logStream.end();
-  }
-};
-
-export const generateManifestForReactNative = async (ctx: Context) => {
-  if (!ctx.isReactNativeApp) return;
-
-  ctx.log.debug('Generating manifest.json file for React Native build');
-  return await generateManifest(ctx);
-};
-
-const maybeTransitionToPendingManifest = (ctx: Context, task: Task) => {
-  if (!ctx.isReactNativeApp) return;
-  transitionTo(pendingManifest)(ctx, task);
-};
-
 /**
  * Sets up the Listr task for building the user's Storybook or E2E project.
  *
@@ -363,15 +227,7 @@ export default function main(ctx: Context) {
     title: initial(ctx).title,
     skip: async (ctx) => {
       if (ctx.skip) return true;
-      if (ctx.isReactNativeApp && ctx.options.storybookBuildDir) {
-        ctx.sourceDir = ctx.options.storybookBuildDir;
-
-        // Use manifest.json from the storybook build directory if it exists
-        if (existsSync(path.resolve(ctx.options.storybookBuildDir, 'manifest.json'))) {
-          return skippedForReactNative(ctx).output;
-        }
-        return false;
-      }
+      if (ctx.isReactNativeApp) return true;
       if (ctx.options.storybookBuildDir) {
         ctx.sourceDir = ctx.options.storybookBuildDir;
         return skipped(ctx).output;
@@ -384,9 +240,6 @@ export default function main(ctx: Context) {
       transitionTo(pending),
       startActivity,
       buildStorybook,
-      buildReactNativeArtifacts,
-      maybeTransitionToPendingManifest,
-      generateManifestForReactNative,
       endActivity,
       transitionTo(success, true),
     ],
