+++ node-src/tasks/index.ts
@@ -3,6 +3,7 @@ import Listr from 'listr';
 import { Context } from '../types';
 import auth from './auth';
 import build from './build';
+import buildReactNative from './buildReactNative';
 import gitInfo from './gitInfo';
 import initialize from './initialize';
 import prepare from './prepare';
@@ -20,6 +21,7 @@ export const runUploadBuild = [
   storybookInfo,
   initialize,
   build,
+  buildReactNative,
   prepare,
   upload,
   verify,
