+++ node-src/types.ts
@@ -169,6 +169,7 @@ export type TaskName =
   | 'storybookInfo'
   | 'initialize'
   | 'build'
+  | 'buildReactNative'
   | 'prepare'
   | 'upload'
   | 'verify'
