+++ node-src/ui/tasks/build.stories.ts
@@ -1,15 +1,13 @@
 import task from '../components/task';
+import { failed, initial, pending, skipped, success } from './build';
 import {
-  failed,
-  initial,
-  pending,
+  failed as failedReactNative,
   pendingAndroid,
   pendingIOS,
   pendingManifest,
-  skipped,
-  skippedForReactNative,
-  success,
-} from './build';
+  skipped as skippedReactNative,
+  success as successReactNative,
+} from './buildReactNative';
 
 export default {
   title: 'CLI/Tasks/Build',
@@ -24,8 +22,6 @@ export const Initial = () => initial(ctx);
 
 export const Building = () => pending({ ...ctx, buildCommand } as any);
 
-export const BuildingReactNative = () => pending({ ...ctx, isReactNativeApp: true } as any);
-
 export const Built = () =>
   success({
     ...ctx,
@@ -34,53 +30,44 @@ export const Built = () =>
     buildLogFile: '/users/me/project/build-storybook.log',
   } as any);
 
-export const BuiltReactNative = () =>
-  success({
-    ...ctx,
-    now: 0,
-    startedAt: -32_100,
-    options: { storybookBuildDir: '/users/me/project/storybook-static' },
-    isReactNativeApp: true,
-    reactNativeBuildLogFile: '/users/me/project/storybook-static/.chromatic/react-native-build.log',
-  } as any);
-
 export const Skipped = () =>
   skipped({
     ...ctx,
     options: { ...ctx.options, storybookBuildDir: '/users/me/project/storybook-static' },
   } as any);
 
-export const SkippedForReactNative = () =>
-  skippedForReactNative({
+export const Failed = () => failed({ ...ctx, buildCommand } as any);
+
+export const BuiltReactNative = () =>
+  successReactNative({
     ...ctx,
-    isReactNativeApp: true,
+    now: 0,
+    startedAt: -32_100,
+    reactNativeBuildLogFile: '/users/me/project/storybook-static/.chromatic/react-native-build.log',
   } as any);
 
-export const BuildingAndroid = () => pendingAndroid({ ...ctx, isReactNativeApp: true } as any);
+export const SkippedForReactNative = () => skippedReactNative();
+
+export const BuildingAndroid = () => pendingAndroid({ ...ctx } as any);
 
 export const BuildingAndroidWithCommand = () =>
   pendingAndroid({
     ...ctx,
-    isReactNativeApp: true,
     options: { reactNative: { androidBuildCommand: 'my-android-build' } },
   } as any);
 
-export const BuildingIOS = () => pendingIOS({ ...ctx, isReactNativeApp: true } as any);
+export const BuildingIOS = () => pendingIOS({ ...ctx } as any);
 
 export const BuildingIOSWithCommand = () =>
   pendingIOS({
     ...ctx,
-    isReactNativeApp: true,
     options: { reactNative: { iosBuildCommand: 'my-ios-build' } },
   } as any);
 
 export const GeneratingManifest = () => pendingManifest();
 
-export const Failed = () => failed({ ...ctx, buildCommand } as any);
-
 export const FailedReactNative = () =>
-  failed({
+  failedReactNative({
     ...ctx,
-    isReactNativeApp: true,
     reactNativeBuildLogFile: '/path/to/project/storybook-static/.chromatic/react-native-build.log',
   } as any);
