+++ node-src/ui/tasks/buildReactNative.stories.ts
@@ -1,0 +1,56 @@
+import task from '../components/task';
+import {
+  failed,
+  initial,
+  pending,
+  pendingAndroid,
+  pendingIOS,
+  pendingManifest,
+  skipped,
+  success,
+} from './buildReactNative';
+
+export default {
+  title: 'CLI/Tasks/BuildReactNative',
+  decorators: [(storyFunction) => task(storyFunction())],
+};
+
+const ctx = { options: {} } as any;
+
+export const Initial = () => initial();
+
+export const Building = () => pending();
+
+export const BuildingAndroid = () => pendingAndroid(ctx);
+
+export const BuildingAndroidWithCommand = () =>
+  pendingAndroid({
+    ...ctx,
+    options: { reactNative: { androidBuildCommand: 'my-android-build' } },
+  } as any);
+
+export const BuildingIOS = () => pendingIOS(ctx);
+
+export const BuildingIOSWithCommand = () =>
+  pendingIOS({
+    ...ctx,
+    options: { reactNative: { iosBuildCommand: 'my-ios-build' } },
+  } as any);
+
+export const GeneratingManifest = () => pendingManifest();
+
+export const Built = () =>
+  success({
+    ...ctx,
+    now: 0,
+    startedAt: -32_100,
+    reactNativeBuildLogFile: '/users/me/project/storybook-static/.chromatic/react-native-build.log',
+  } as any);
+
+export const Skipped = () => skipped();
+
+export const Failed = () =>
+  failed({
+    ...ctx,
+    reactNativeBuildLogFile: '/users/me/project/storybook-static/.chromatic/react-native-build.log',
+  } as any);
