+++ node-src/ui/tasks/utilities.ts
@@ -3,7 +3,6 @@ import { Context } from '../../types';
 
 export const buildType = (ctx: Context) => {
   if (isE2EBuild(ctx.options)) return 'test suite';
-  if (ctx.isReactNativeApp) return 'React Native Storybook';
   return 'Storybook';
 };
 export const capitalize = (string: string) => string.charAt(0).toUpperCase() + string.slice(1);
